mongo

+++ rebuild.bat
@@ -7,19 +7,19 @@ docker-compose down -v
 echo.
 
 echo docker images remove...
-FOR /F "tokens=1" %%i IN ('docker images -q --filter "dangling=false"') DO (
+FOR /F "tokens=*" %%i IN ('docker images -q --filter "dangling=false"') DO (
     docker rmi %%i
 )
 echo.
 
 echo folder remove...
-rmdir /s /q .\fastapi\logs
-rmdir /s /q .\mysql\data
-rmdir /s /q .\mysql\logs
-rmdir /s /q .\mongo\data
-rmdir /s /q .\mongo\log
+IF EXIST .\fastapi\logs rmdir /s /q .\fastapi\logs
+IF EXIST .\mysql\data rmdir /s /q .\mysql\data
+IF EXIST .\mysql\logs rmdir /s /q .\mysql\logs
+IF EXIST .\mongo\data rmdir /s /q .\mongo\data
+IF EXIST .\mongo\log rmdir /s /q .\mongo\log
 echo.
 
 echo Docker Compose on...
-docker-compose up
+docker-compose up -d
 ENDLOCAL
